rfgo-rag-backend/pyproject.toml
[project]
name = "rfgo-rag-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "fastmcp[cli]>=3.2.0",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.4",
    "openai>=2.30.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
